scripts/tues: Delete uploaded files after command execution
We previously left behind the files provided through the `-f` switch.
To fix this, move the file handling code into a context manager. We
now use the return value of `fabric.api.put` to determine whether the
upload succeeded and to provide the absolute path to the uploaded file
in the environment variable.