Reimplement with asyncssh
Since the fabric interface has changed quite a bit, it's no longer
an option for us. It was nice not having to do much and get the
sudo magic for free, but on the other hand implementing it manually
gives us the freedom to mix in some features from our own wishlist
which were pretty much impossible to do while using fabric.

Additionally, the code has been prepared to be used as a library
as much as possible, leaning towards a `subprocess.run`-ish interface.

In the process of reimplementing the commandline interface, the `-p`
option to trigger parallel execution has been deprecated.

Concurrency is now solely triggerd by setting `-n <num>`. The compat
option `-p` will simply set `-n 10` for now and will be removed in a
future release.

The `--output-command` option has also been removed, it was a bit of
a "because I can" feature that I am now thinking should be implemented
in a completely different way, so we use the major version upgrade to
get rid of it, only to introduce something similar again later..maybe..

There is also now "script support". You may define a `TUES_PATH`
variable (defaulting to `$XDG_CONFIG_HOME/tues/scripts/"), to store
executable files that will be copied to the remote host before
execution, the variable is subject to the same mechanis as the `PATH`
variable and can contain multiple directories seperated by a `:`.
The option itself is just a flag, the usual `cmd` is
used for the name of the script. The code can handle `cmd` containing
additional options instead of just the name of the script, in which
case it will first copy the script and then call it with those options.

The "sudo handling" has been improved (hopefully) to not leave any
traces of a sudo prompt (or "reprompt") including any superfluous
newlines that sudo might emit.

While handling newlines, we are not also now don't emit a line prefix
if the command does not generate *any* output.

To satisfy my own ocd, I have also added an option that will pad the
prefix with spaces on the left in order to align the output of all
hosts on the same column.