Make TUES_FILE<x> contain absolute paths
This is even more important now that we have the `cwd` argument to run.
Even previously, people could very well be `cd`-ing into some other
directory and it is just inconvenient to have to keep track of stuff
like this manually.

The implementation is a bit "meh" though. The old `fabric.put` operation
already returned absolute paths of all uploaded files, but neither the
scp implementation we currently use, nor the sftp alternative does that.

What I ended up doing is move the `cd {¢wd}` portion of the `precmds`
as close as possible to the actually executed command. This way the
current working directory during envvar evaluation is the home dir
of the login user, which is also where we upload the files to.

Then just using $PWD"/filename" will expand to the absolute path of
the file. While not very pretty, it seems to work in most cases and
is not dependent on $HOME pointing to the original users home instead
of the sudo users'.