Make output-dir handling configurable
In order to make it easier for people to keep a history of runs or just
to be sure the output dir only contains output from a single run, we
introduce "strategies" to deal with an existing output_dir.

There is a new argument `output_dir_strategy` for `run` as well as
a new `-S/--output-dir_strategy` option for the CLI.

It accepts 4 values `DIR_IGNORE`, `DIR_ABORT`, `DIR_WIPE` and `DIR_ROTATE`.

  * Ignore will just use the directory as-is, potentially leaving
    a mix of files from a previous run, if they are not part of the
    current run. Existing files will be truncated, though.

  * Abort will always abort when the directory exists, without
    regard for its contents.

  * Wipe will remove all "*.log" files from the directory before
    running any commands.

  * Rotate will collect all <dir>.<integer> directories, figure out
    and increment the highest integer found, and rename the old <dir>
    to that new name and create a new <dir> for the current run.