Try to terminate the remote command on ctrl-c
If the user hits ctrl-c, we do our best to terminate the remote command
by either sending it the `TERM` signal through the ssh server, this has
the drawback of only working with more recent openssh servers, because
a bug (opened in 2008) was only fixed in 2019, so for example only
Debian Buster has a recent enough ssh server.

The good news is that with `pty=True`, we can send a special control
sequence on stdin, which will result in the process receiving `SIGINT`.

Additionally, when changing users via sudo, we cannot send signals
directly, since we do not have proper permissions, we need to rely
on the "pty method". For that reason we also change the default
of the CLI to enable pty mode.

The Python API still defaults to `pty=False`, since we expect the
Python API to be used in a much more "non-interactive" way while also
being used in a way that is not as prone to random commands blocking
on stdin (like apt-get install etc.).

I checked other tools like fabric and they essentially all do the same.
The more elaborate solution would be to also echo the process group id
of the shell used to wrap out sudo command, which we could easily do
in the expression itself without much problems, and then use that ID
to run `sudo kill -<id>` in the shutdown handler on the new channel
of the same session. For now this seems to be a bit too clever and
much more hassle than it is worth. Perhaps time will prove me wrong. :-)