Handle connection errors, even with concurrency
We now properly wrap connection errors like timeouts, lookup errors etc.
into a `TuesError`. When running with concurrency, we will collect
all exceptions and results and if there really are exceptions, we will
wrap them in a `TuesErrorGroup`, containing all the exceptions on the
`.exceptions` property and all completed `Task` instances on the
`.results` property for inspection by the user.

We also display all encountered connection errors during execution at
the end of the output so they don't scroll out of the view of the
caller.