Allow passing ports in host string or as (host, port) tuple
In order to be able to connect to a mix of ports, we need to get the
port from each `host` individually (it it is a list). Since we need
to handle the host (and retrieve the portless hostname) in multiple
places, we add a `Host` class that wraps the parsing logic instead
of just doing everything in the `Task` class.

We use `urllib.parse.urlparse` with a "ssh://" prefix to parse the
port from a "addr:port" string. This will give us proper host, ipv4
and ipv6 parsing almost for free. :-)