Allow/Require message argument in PasswordManager.get
We will be adding an option to enter a login password on the commandline.
Since it will be the same password as the sudo password, we should use
the same `PasswordManager` instance, so we need a way to display individual
prompt messages on the same instance, passing an argument to `get` seems
the most sensible option. Having a "default" message seems to make not
much sense though, so drop it from the constructor and make it mandatory
in `get`.