Allow login with password auth
Reuse the `PasswordManager` instance since the login password will match
the sudo password (I guess there are some crazy sudo setups where this
assumption may be false, but we'll them for now).

Also, doing this will probably not be a smart idea in cases where you
have a lot of different passwords in use, most notably in tandem with
something like fail2ban. :-D

Also, you don't want to use this in parallel mode when you are not sure
your password will be accepted on all hosts while also not redirecing
command output to a directory, you probably won't even see the prompt
being displayed and we certainly won't add an option to halt output
just to prompt for a password, this is purely an emergency option that
should not be used on the regular.