Allow managing newline transformation in pty mode
When in pty mode, `\n` is automatically transformed to `\r\n`. While
this was or may still be necessary when sending a bytestream directly
to a pty. Since we are running in pty mode by default (a choice I more
and more regret every day), we are now having our output automatically
transformed, even if it is a bytestream that does not do well with such
intervention, like running `tcpdump` on the remote side and streaming
into into a local wireshark instance...

Since this behaviour is "just" the default and can be reset on the
fly by `stty` or from within tues with asyncsshs `term_modes`, we
introduce the `universal_newlines` flag to control the behaviour.

When using tues as a library, we strictly sick to the defaults, and
just make the commandline `tues` prevent pty mode from automatically
transforming the newlines, thus keeping the default behaviour for
most of the code and moving the "special bits" up to the cli which
will probably be the only user of this feature anyway.