Allow disabling line prefixing from script run args
Usually the script that is producing the output knows best if a script
is run in a context where line prefixes are useful or not, so we let
the script decide whats best. Since the code layout "forces" us to
pass an explicit value from the cli to `tues.run` and we'll just
disregard that value in script mode if the script passes the proper
key in its args, we'll have to live with not being able to force a
prefix from the commandline.