Fix password re-prompting in parallel mode
When running multiple tasks in parallel with a bad password, every
task while invalidate the fixed password in turn and trigger another
reprompt, until eventually all tasks have prompted again.

Since I do not have a nice lock based solution at the moment - the
code is a bit to async-y for that I fear. We go another route and
just "guess" if someone else already did the reprompt by remembering
the last password we attempted and only invalidate if we already
know the password to be bad.

Since the `SudoWriter` is the first item with "per task" data, we
are currently using it to also keep the password around, making
this solution purely caller based, which is not as nice as I would
like the interface of `PasswordManager` to be.

For now this will have to do though, because I still can't shake the
feeling we want a better solution and I don't want to needlessly
change the interface again and again.