Switch executables to console entrypoints
I have resisted this for quite some time, but I guess there is no
denying that this is the standard we should follow now, everything
else is just not going to be supported anymore. Even though I do
not like having these wrapper scripts around the real code, it will
be more trouble that it is worth to keep this around in the future.