Move code to src/
The cool kids seem to be doing this a lot these days, so we might also
dip our toe into the waters. Since starting to code in Rust a bit, I
have already adapted a bit to it, so we might as well try this out for
Python, too. I am not quite sure if we should also move over the tests,
but I could not get a clear picture on which way people tend to lean
on this. So for now we keep the test/ directory in the root.