Do not ignore cli params if the script does not use them
We ignore certain variables/options like `prefix` when reading config
from a script, because if the script has set the variable, it might have
a good reason to do so. We should not blindly drop the value from the
CLI though, just because the variable is in the list of "ignorables".
Only if the variable is known to be part of the scripts run args should
the CLI value be ignored.