Make `run` return type more consistent
We previously returned a single `Task` instance (instead of a list)
whenever we handled only one host. That is suprising to the caller,
though, when working with a dynamic list of hosts that may or may not
contain more than one element. So make this more consistent by always
matching the type of the `hosts` parameter (i.e. always return a list
when we are given a list, regardless of the length).