Rewrite parallelization and cancellation handling
The previous approach to handling SIGINT was insufficient because we
handled the signal _outside_ the loop which meant that we could not
really interact with the SSH session any more. This had the effect
that we just closed/aborted the session without terminating the remote
process.

This rewrite uses the signal-handler feature of the asyncio loop. The
signal handler is executed from within the loop and can cancel our
tasks. Cancelling a task simply raises an exception from the current
`await` site. This exception can be handled, which allows us to
properly terminate the session and clean up after ourselves.

Rewriting of the parallelization isn't strictly required for this, but
happened in the process (and I'm not going to try to pick this apart).
The previous implementation started all per-host tasks at once and
then had them block on a semaphore as needed to stay within the pool
size limit. The new implementation only starts enough tasks to keep
the pool "full", then waits for some to finish and starts new ones.
This should be the smarter approach with large host sets.

For further work, it might be a good idea to remove the special casing
of "pool_size = 1" to reduce the amount of code we have. But that
requires integrating the "check" feature into the parallel case.