Try to fix env passing
There are two issues at play here. First, passing environment
assignments to `create_session` is mostly useless because the SSH
server only allows setting very few variables. We now pass the user
provided environment there for the case that the user actually wants
to set one of the allowed variables. The `TUES_FILE<n>` variables we
generate ourselves are unlikely to be allowed.

Second, we previously generated simple `VAR="value"` assignments as
pre-commands. This was enough to pass the `echo $VAR` test, but
doesn't actually add the variables to the environment of the processes
we run. So use `export` and test for this by running `env` in the
test.