Try to disentangle `Run` and `SudoWriter` objects
Rather than having the `SudoWriter` inject a pre-command and wrapper
on the run, let the main run logic ask the writer to wrap the command.
This allows us to not pass the run to the writer.

Update the writer's docstring while we're at it, this seemed outdated
to begin with.