Refactor sudo setup
Massage the code a bit to the point where we can pass the "back
channel" of the `SudoWriter` to the constructor rather than assigning
it later on. This clarifies the lifecycle of the class because it is
clearly not usable without the back channel set. While we're at this,
make the back channel property "private" and rename it so we don't
have a "sudo" property on a "sudo" object any more.

We can also drop the creation of the `OutputWrapper` for the sudo
writer object because we never used it.