Refactor sudo setup some more
We have now massaged everything to the point where we can pass the
sudo "back channel" (where the `SudoWriter` needs to write the
password, i.e. the stdin channel of the session) directly to the
`SudoWriter` constructor instead of assigning it later on. This makes
the lifecycle of the `SudoWriter` class a little clearer since it
clearly cannot work without the back channel.

While we're at this, make the back channel property "private" and
rename it so we don't have a "sudo" property on a "sudo" object any
more.