Support aborting sudo password entry via Ctrl+C
This ability was lost when we switched to asyncio-aware sigint
handling via `loop.add_signal_handler`. In that mode, signals are only
handled during async task switches, which never happens when the
current task is blocked reading user input. We solve this primary
issue by temporarily restoring the regular Python sigint handler (and
disabling the wakeup-fd registered by the event loop).

The problem is further complicated by the fact that the password
prompt is triggered from the `write` function of the `SudoWrapper`.
That function is invoked by a background task created by asyncssh and
is thus not part of the (await) call stack of our async task. This
means that cancelling our main tasks does not raise an exception
within that background task.

We solve this by making the `write` function close its back channel
(stdin of the remote `sudo` process) when the user cancels the
password prompt. This terminates the remote process which in turn
tears everything down cleanly.

Finally, we have to transport the information that the user did not
authenticate back to the main task and then raise an exception to our
caller.