Rewrite cancellation handling once more
We previously installed our custom signal handlers in the low level
run functions. Loop signal handlers are (loop-) global, though, and
replace existing handlers, which is not nice behavior for a library.
It turns out that we can push up the signal handler into our top-level
`run` function by pre-creating an asyncio task and passing that
(instead of the coroutine) to `run_until_complete`. This means that
the low-level run functions only have to deal with the cancellation
exception and cancel the tasks they created.

We still do mess with signal handling to some degree, especially if a
caller passes an existing loop to `run`. We should probably drop that
feature.