Fix unbound local error
The `cleanup` callback list is used in the `finally` branch, but may
not be set, yes, when an exception happens too early in the `try`
block.

There may be a more elegant solution for this hiding somewhere, but
I'd like to get a fix out before setting out to find it.