Preserve exception context when raising `TuesTaskError`
We currently raise `TuesTaskError` both when the task executed but
returned non-zero (when `check` is set) and when there was a problem
executing the task. In the latter case, the task has no returncode
set, which makes the exception not-very-helpful. Preserving the
context of the underlying exception should help.

Note that the `raise ... from exc` is valid even when `exc` is `None`.
This is usually used to suppress the context when raising from an
`except` clause. In our case, it doesn't have any effect because there
is no context to suppress, but works nonetheless.