Add --debug to optionally show tracebacks
We usually catch all exceptions that bubble up to the toplevel and wrap
them into a ClickException to make them look a bit nicer, but sometimes
that is detrimental to debugging simple errors because we cannot see the
traceback anymore while the error is pretty generic. So let us at least
have an option to provide any user with an easy way to get that.