Add asyncssh version constraint
We are using `create_session(recv_eof)`, which is not supported in
versions prior to 2.13, so we should demand at least that version.
We are currently at 2.15, but I hesitate to pull it all the way up
because we might be mixed into another application and don't want
too (as of now) needlessly constrain pips wiggle room.