Fix debug switch being an integer
There is always this mental reflex that an option is recognized as a
flag, just by giving a boolen default value, which is obviusly wrong.
Nontheless, I usually cannot help myself and fall into the same trap
again. No idea how I could test this with it being buggy, but I did
actually not only try this out but introduced this explicitly for
debugging something, so it must have done something. I probably was
not paying attention to the details, because I was debugging something
else though.

Well, not there is at least one very minimal test...