Guard against negative / None returncode
The returncode can be negative when exiting from a signal, and it may
even be `None` when there are some error during communication before
the execute command can exit properly.

To be able to better debug the source of the `None` values, we first try
getting rid of any followup exceptions that might shadow the source of
these problems.