cli/tues: Error out when we find no hosts
Error out when a provider is given on the command line but does not
find hosts. This was already handled in 53b8596f99 (tues: Handle "no
hosts" case in `run`, 2023-07-20) for runs with an explicit command
line, but still failed badly in the `tues -s <script>` case because
the `Script` class would try to find the hosts itself (the provider
settings _can_ be given in the script), but crashed when the script
did not specify a provider and/or provider args.

While the various supported use cases make this all a bit complex, it
seems safe to say that _if_ a provider is given on the command line,
we should abort if it does not find hosts. Even if the script in
question does specify a provider call, we should not fall back to it
when an explicit one was given.