Adapt to changes in asyncssh 2.19
The 2.19 release introduced some new mandadatory arguments to the
interfaces of the ssh config classes. Unfortunately, the `load`
classmethod has a particularly inconvenient interface that does not
allow us to use keyword arguments. Do the next best thing and comment
all the parameters:

NOTE: I'm not sure the values for the new `canonical` and `final`
      parameters are correct; the whole change is quite scarce on
      documentation. But I think this only affects conditional
      sections in the config file.